use logger in action-logger also. move objToJS to periodic-logger
